Created the Object, Section, Symbol and Relocation classes.
Their purpose is to read and relate ELF Object Files' contents.

Also added the Map class, that implements a map-like container on
top of a sorted vector, as recommended by LLVM as an efficient
map for several purposes, different from std::map.

These changes kind of broke the SBT for now (although it was not
really functional before either...).
There are two main missing pieces that need to be done to fix this:
1- Relate the Symbols to their corresponding Sections, ordered by
their addresses
2- Finish the code that deals with relocations