The SBT is now correctly translating a HelloWorld binary.
The SBT can now translate a HelloWorld RISC-V object file, generate
from sbt/test/hello.s source code.
The source does only 2 syscalls: write and exit.
However, it already needs to deal with data and relocations in
order to be able to pass a valid pointer to the syscall.

Current status of the SBT:
  Instructions translated:
  - addi
  - auipc
  - ecall
  Features:
  - can handle data section relocations
  - generates a shadow image of the data section
  - supports register file using global variables
  TODO (near future):
  - add several other instructions
  - support stack
  - support other types of relocations
  - handle calls to other modules
  - improve symbol and function handling
  - translate object files generated from C code