Updated docker build.
Building with success and passing in all tests.

The last image uses multistage docker build, to greatly reduce the size
of the image. Ideally, the intermediate images should use this techique too,
to reduce their size, specially because removing source files after the build
doesn't reduce the size of the image, because of Docker's overlay FS. The
solution is to perform a 2 stage build for each image, so that the second stage
can copy only what is needed from the first stage and then discard it.