Bug fix in register synchronization
There was a wrong optimization in Function::storeRegisters(), in which
local registers that were never written were not being stored to their
global equivalents.
But this is incorrect, because control flow is not considered when
translating code, and thus we may not assume the register were never
written before, because we don't know the paths the code took to
reach the current translation point.

So, for instance, a jump forward, followed by a register write,
followed by a jump backward could cause a register that needs to be
synchronized for correct operation. Ex:

        # no writes to fp
        jmp L2
L1:
        # reg sync
        call adpcm_coder
        # reg sync
        mov [mem], fp
        jmp L3
L2:
        mov fp, 2000
        jmp L1
L3:
        ...

This was more or less what was happening in adpcm-encode.
Reg sync was incorrectly skipping synchronization of fp register,
because, sequentially, it was never written before, but, considering
the flow path, it actually was written before the call to adpcm_coder
and thus should be saved before the call.
The conclusion is that the translator can't make any assumption on
previous register writes, at least not considering only the code it
translated before.