Improved icall translation
On some benchmarks, such as BitCount, being able to assume that all
indirect calls are to internal functions only can improve translation
quality substantiatly, by eliminating the check-external-branch part.
To this end, a new -icall-int-only option was introduced in the SBT,
telling it to assume all icalls are to internal functions.
This has a 20% reduction on ARM slowdown on BitCount benchmark in ABI
mode.

Besides that, there were also some minor changes and notes added due to
blowfish and bitcount performance analysis.